systest/requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
python-dotenv>=0.19.0
requests>=2.25.0
unittest2>=1.1.0